Remove main.go -> cmd/ministries/main.go
