+++ go.mod
@@ -10,7 +10,7 @@ require (
 	github.com/stretchr/testify v1.4.0 // indirect
 	go.uber.org/zap v1.10.0
 	golang.org/x/sys v0.0.0-20200116001909-b77594299b42 // indirect
-	gopkg.in/natefinch/lumberjack.v2 v2.0.0
+	gopkg.in/natefinch/lumberjack.v2 v2.0.0 // indirect
 	gopkg.in/sourcemap.v1 v1.0.5 // indirect
 	gopkg.in/yaml.v2 v2.2.8 // indirect
